feat(checkout): CHECKOUT-10152 add new useBilling hook to manage common logic between 2 billing components
