+++ packages/core/src/app/billing/Billing.tsx
@@ -1,19 +1,13 @@
-import type { CheckoutSelectors, FormField } from '@bigcommerce/checkout-sdk';
-import React, { type ReactElement, useCallback, useEffect, useState } from 'react';
+import type { CheckoutSelectors } from '@bigcommerce/checkout-sdk';
+import React, { type ReactElement } from 'react';
 
-import { useCapabilities, useCheckout } from '@bigcommerce/checkout/contexts';
 import { TranslatedString } from '@bigcommerce/checkout/locale';
 import { AddressFormSkeleton, Legend } from '@bigcommerce/checkout/ui';
 
-import {
-    AddressType,
-    isEqualAddress,
-    mapAddressFromFormValues,
-    setDefaultAddress,
-} from '../address';
+import { isEqualAddress, mapAddressFromFormValues } from '../address';
 
 import BillingForm, { type BillingFormValues } from './BillingForm';
-import getBillingMethodId from './getBillingMethodId';
+import { useBilling } from './hooks/useBilling';
 
 export interface BillingProps {
     navigateNextStep(): void;
@@ -21,76 +15,18 @@ export interface BillingProps {
     onUnhandledError(error: Error): void;
 }
 
-const getFieldsWithExtraFields = (
-    getBillingAddressFields: (countryCode: string) => FormField[],
-    hasAddressExtraFields: boolean,
-    getAddressExtraFields: () => FormField[],
-    countryCode?: string,
-) => {
-    const addressFields = getBillingAddressFields(countryCode || '');
-
-    if (!hasAddressExtraFields) {
-        return addressFields;
-    }
-
-    const addressExtraFields = getAddressExtraFields();
-
-    return [...addressFields, ...addressExtraFields];
-};
-
 const Billing = ({ navigateNextStep, onReady, onUnhandledError }: BillingProps): ReactElement => {
     const {
-        selectedState: {
-            checkout,
-            config,
-            cart,
-            customer,
-            isLoadingBillingCountries,
-            getBillingAddressFields,
-            getAddressExtraFields,
-        },
-        // using getBillingAddress function to guarantee latest state inside of async function in useEffect
-        checkoutState: {
-            data: { getBillingAddress },
-        },
-        checkoutService,
-    } = useCheckout(({ data, statuses }) => ({
-        checkout: data.getCheckout(),
-        config: data.getConfig(),
-        cart: data.getCart(),
-        customer: data.getCustomer(),
-        billingAddress: data.getBillingAddress(),
-        isLoadingBillingCountries: statuses.isLoadingBillingCountries(),
-        getBillingAddressFields: data.getBillingAddressFields,
-        getAddressExtraFields: data.getAddressExtraFields,
-    }));
-    const {
-        userJourney: { hasAddressExtraFields, hasCompanyAddressBook },
-        billing: { restrictManualAddressEntry },
-    } = useCapabilities();
-
-    if (!config || !customer || !checkout || !cart) {
-        throw new Error('Unable to access checkout data');
-    }
-
-    const [isApplyingDefaultAddress, setIsApplyingDefaultAddress] = useState(true);
-    const isInitializing = isLoadingBillingCountries || isApplyingDefaultAddress;
-
-    // Below constants are for <BillingForm />'s HOC props
-    const customerMessage = checkout.customerMessage;
-    const methodId = getBillingMethodId(checkout);
-    const billingAddress = getBillingAddress();
-
-    const getFields = useCallback(
-        (countryCode?: string) =>
-            getFieldsWithExtraFields(
-                getBillingAddressFields,
-                hasAddressExtraFields,
-                getAddressExtraFields,
-                countryCode,
-            ),
-        [getBillingAddressFields, hasAddressExtraFields, getAddressExtraFields],
-    );
+        billingAddress,
+        customerMessage,
+        getBillingAddress,
+        getFields,
+        isInitializing,
+        methodId,
+        showNoAddressesWarning,
+        updateBillingAddress,
+        updateCheckout,
+    } = useBilling({ onReady, onUnhandledError });
 
     const handleSubmit = async ({
         orderComment,
@@ -101,11 +37,11 @@ const Billing = ({ navigateNextStep, onReady, onUnhandledError }: BillingProps):
         const address = mapAddressFromFormValues(addressValues);
 
         if (address && !isEqualAddress(address, billingAddress)) {
-            promises.push(checkoutService.updateBillingAddress(address));
+            promises.push(updateBillingAddress(address));
         }
 
         if (customerMessage !== orderComment) {
-            promises.push(checkoutService.updateCheckout({ customerMessage: orderComment }));
+            promises.push(updateCheckout({ customerMessage: orderComment }));
         }
 
         try {
@@ -119,38 +55,7 @@ const Billing = ({ navigateNextStep, onReady, onUnhandledError }: BillingProps):
         }
     };
 
-    useEffect(() => {
-        const init = async () => {
-            try {
-                await checkoutService.loadBillingAddressFields();
-
-                if (hasCompanyAddressBook) {
-                    await setDefaultAddress({
-                        type: AddressType.Billing,
-                        currentAddress: getBillingAddress(),
-                        addresses: customer.addresses,
-                        updateAddress: checkoutService.updateBillingAddress,
-                    });
-                }
-
-                onReady();
-            } catch (error) {
-                if (error instanceof Error) {
-                    onUnhandledError(error);
-                }
-            } finally {
-                setIsApplyingDefaultAddress(false);
-            }
-        };
-
-        void init();
-    }, []);
-
-    // Show warning message when restrictManualAddressEntry is true and no addresses are available
-    const hasAddresses = customer?.addresses && customer.addresses.length > 0;
-    const showWarningMessage = restrictManualAddressEntry && !hasAddresses;
-
-    if (showWarningMessage) {
+    if (showNoAddressesWarning) {
         return (
             <div className="no-addresses-warning body-regular">
                 <TranslatedString id="billing.no_billing_addresses_warning" />
